requirements.txt
whois
request
subdomain_scanner
colorama
python-whois
